+++ app/Http/Controllers/FournisseurController.php
@@ -11,54 +11,78 @@ class FournisseurController extends Controller
 {
     public function add(Request $request) {
 
-        $this->validate(
-            $request,
-            ['nom_fournisseur' => 'required|min:5'],
-            ['nom_fournisseur.required' => 'Le nom est obligatoire!'],
-            ['telephone' => 'required|numeric|regex:/(0)([5-7])[0-9]{8}/'],
-            ['telephone.required' => 'Telephone est obligatoire!']
-        );
+        if (Auth::check()) {
 
-        $fournisseur = new Fournisseur;
-        $fournisseur->nom = $request->nom_fournisseur;
-        $fournisseur->telephone = $request->telephone;
-        $fournisseur->activite = $request->activite;
-        $fournisseur->region = $request->region;
+            $this->validate(
+                $request,
+                ['nom_fournisseur' => 'required|min:5'],
+                ['nom_fournisseur.required' => 'Le nom est obligatoire!'],
+                ['telephone' => 'required|numeric|regex:/(0)([5-7])[0-9]{8}/'],
+                ['telephone.required' => 'Telephone est obligatoire!']
+            );
 
-        $user = Auth::user();
+            $fournisseur = new Fournisseur;
+            $fournisseur->nom = $request->nom_fournisseur;
+            $fournisseur->telephone = $request->telephone;
+            $fournisseur->activite = $request->activite;
+            $fournisseur->region = $request->region;
 
-        $user->fournisseurs()->save($fournisseur);
+            $user = Auth::user();
 
-        return back();
+            $user->fournisseurs()->save($fournisseur);
+
+            return back();
+
+        } else {
+            return redirect('/login');
+        }
     }
 
     public function delete(Fournisseur $fournisseur) {
 
-        $fournisseur->delete();
+        if (Auth::check()) {
 
-        return back();
+            $fournisseur->delete();
+
+            return back();
+
+        } else {
+            return redirect('/login');
+        }
 
     }
 
     public function edit(Fournisseur $fournisseur) {
 
-        return view('fournisseurs.editPage', compact('fournisseur'));
+        if (Auth::check()) {
+
+            return view('fournisseurs.editPage', compact('fournisseur'));
+
+        } else {
+            return redirect('/login');
+        }
 
     }
 
     public function update(Request $request, Fournisseur $fournisseur) {
+        
+        if (Auth::check()) {
 
-        $this->validate(
-            $request,
-            ['nom' => 'required|min:5'],
-            ['nom.required' => 'Le nom est obligatoire!'],
-            ['telephone' => 'required|numeric|regex:/(0)([5-7])[0-9]{8}/'],
-            ['telephone.required' => 'Telephone est obligatoire!']
-        );
+            $this->validate(
+                $request,
+                ['nom' => 'required|min:5'],
+                ['nom.required' => 'Le nom est obligatoire!'],
+                ['telephone' => 'required|numeric|regex:/(0)([5-7])[0-9]{8}/'],
+                ['telephone.required' => 'Telephone est obligatoire!']
+            );
 
-        $fournisseur->update($request->all());
+            $fournisseur->update($request->all());
 
-        return redirect('programmer');
+            return redirect('/programmer');
+
+        } else {
+            return redirect('/login');
+        }
 
     }
 }
